Search page loading no filter but yes uni info. Changed SA course to include uni information

+++ backend/server/src/main/java/model/ICourse.java
@@ -3,5 +3,5 @@ package model;
 public interface ICourse {
   String getCourseNumber();
   String getCourseTitle();
-  int getCredits();
+  double getCredits();
 }
